// src/common/enum/error-code.enum.ts
export enum ErrorCode {
  TOKEN_EXPIRED = 'TOKEN_EXPIRED',
  INVALID_TOKEN = 'INVALID_TOKEN',
  TOKEN_NOT_PROVIDED = 'TOKEN_NOT_PROVIDED',
  UNAUTHORIZED_CONCERT_PAYMENT = 'UNAUTHORIZED_CONCERT_PAYMENT',
  PAYMENT_ALREADY_PROCESSED = 'PAYMENT_ALREADY_PROCESSED',
  PAYMENT_EXPIRED = 'PAYMENT_EXPIRED',
  BOOKING_NOT_STARTED = 'BOOKING_NOT_STARTED',
  BOOKING_ALREADY_PASSED = 'BOOKING_ALREADY_PASSED',
  SEAT_ALREADY_BOOKED = 'SEAT_ALREADY_BOOKED',
  SEAT_ALREADY_SOLD = 'SEAT_ALREADY_SOLD',
  SEAT_NOT_RESERVED = 'SEAT_NOT_RESERVED',
  CONCERT_NOT_FOUND = 'CONCERT_NOT_FOUND',
  SCHEDULE_NOT_FOUND = 'SCHEDULE_NOT_FOUND',
  SEAT_NOT_FOUND = 'SEAT_NOT_FOUND',
  BOOKING_NOT_FOUND = 'BOOKING_NOT_FOUND',
  MINIMUM_CHARGE_AMOUNT = 'MINIMUM_CHARGE_AMOUNT',
  MINIMUM_PAYMENT_AMOUNT = 'MINIMUM_PAYMENT_AMOUNT',
  INSUFFICIENT_POINT = 'INSUFFICIENT_POINT',
  POINT_NOT_FOUND = 'POINT_NOT_FOUND',
  QUEUE_NOT_FOUND = 'QUEUE_NOT_FOUND',
  OPTIMISTIC_LOCK_CONFLICT = 'OPTIMISTIC_LOCK_CONFLICT',
  INTERNAL_SERVER_ERROR = 'INTERNAL_SERVER_ERROR',
}